Arima_based/requirements.txt
# RippleNet-TFT Requirements
# Core ML and Data Science
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Time Series Analysis
statsmodels>=0.13.0
pmdarima>=2.0.0

# Graph Neural Networks
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17
torch-cluster>=1.6.1

# Transformers and NLP
transformers>=4.21.0
tokenizers>=0.13.0
datasets>=2.0.0

# Financial Data
yfinance>=0.1.87
requests>=2.28.0

# Sentiment Analysis
vaderSentiment>=3.3.2

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
tqdm>=4.64.0

# Data Processing
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# Optional: Experiment Tracking
wandb>=0.13.0

# Optional: Advanced Time Series
pytorch-forecasting>=0.10.0

# Development and Testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0